# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.UploadAssetResource:
    type: object
  models.AssetResource:
    properties:
      description:
        type: string
      id:
        type: integer
      program:
        $ref: '#/definitions/models.Program'
      program_id:
        type: integer
      resource:
        $ref: '#/definitions/models.Resource'
    type: object
...
  models.FileExtension:
    properties:
      extension_name:
        type: string
      id:
        type: integer
      mime_type:
        type: string
    type: object
  models.PDF:
    properties:
      id:
        type: integer
      pages:
        items:
          $ref: '#/definitions/models.PDFPage'
        type: array
      resource_id:
        type: integer
    type: object
  models.PDFPage:
    properties:
      content:
        type: string
      id:
        type: integer
      page_number:
        type: integer
      pdf_id:
        type: integer
    type: object
  models.Program:
    properties:
      id:
        type: integer
      program_name:
        type: string
    type: object
...
  models.Resource:
    properties:
      created_at:
        type: string
      file_extension:
        $ref: '#/definitions/models.FileExtension'
      file_extension_id:
        type: integer
      id:
        type: integer
      path:
        type: string
      pdf:
        $ref: '#/definitions/models.PDF'
      resource_name:
        type: string
      resource_type:
        $ref: '#/definitions/models.ResourceType'
      resource_type_id:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  models.ResourceType:
    properties:
      id:
        type: integer
      type_name:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /v1/assetResources:
    post:
      description: Upload a new asset resource for a specific program
      parameters:
      - description: Upload Asset Resource Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UploadAssetResource'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AssetResource'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload an asset resource
      tags:
      - Resource
  /v1/assetResources/{id}:
    delete:
      description: Delete an asset resource by its ID
      parameters:
      - description: Asset Resource ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an asset resource
      tags:
      - Resource
  /v1/assetResources/{program_id}:
    get:
      description: Retrieve all asset resources associated with a specific program
      parameters:
      - description: Program ID
        in: path
        name: program_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.AssetResource'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get asset resources by program ID
      tags:
      - Resource
...
  /v1/projectResources/{id}:
    delete:
      description: Delete a project resource and its associated file
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a project resource
      tags:
      - Resource
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"